등록&임시저장&삭제&수정 기본 로직추가
등록&임시저장&삭제&수정 기본 로직추가

+++ src/main/java/com/DOH/DOH/dto/notifications/EventDTO.java
@@ -10,7 +10,7 @@ import java.time.LocalDate;
 @NoArgsConstructor
 @AllArgsConstructor
 public class EventDTO {
-    private int eventNum;
+    private Long eventNum;
     private String eventTitle;
     private String eventContent;
     private LocalDate eventCreateTime;
